+++ CHANGELOG.md
@@ -4,6 +4,10 @@ branch is the latest unreleased version.
 
 ## imperative
 **Highlights:**
+- Add `rbxmk download-asset` command to quickly download an asset from the
+  Roblox website.
+- Add `rbxmk upload-asset` command to quickly upload an asset to the Roblox
+  website.
 - Add `rbxmk dump` command to dump the rbxmk Lua API in various formats.
 	- Supports generic JSON and minified JSON format.
 	- Supports selene TOML format.